Move Lrama::Symbol to Lrama::Grammar::Symbol
