+++ lib/lrama.rb
@@ -12,6 +12,5 @@ require "lrama/report"
 require "lrama/state"
 require "lrama/states"
 require "lrama/states_reporter"
-require "lrama/symbol"
 require "lrama/version"
 require "lrama/warning"
