+++ lib/lrama/grammar.rb
@@ -2,6 +2,7 @@ require "lrama/grammar/code"
 require "lrama/grammar/precedence"
 require "lrama/grammar/printer"
 require "lrama/grammar/rule"
+require "lrama/grammar/symbol"
 require "lrama/grammar/union"
 require "lrama/lexer"
